+++ database/migrations/2016_06_01_000005_create_oauth_personal_access_clients_table.php
@@ -14,12 +14,12 @@ class CreateOauthPersonalAccessClientsTable extends Migration
     public function up()
     {
         Schema::create('oauth_personal_access_clients', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->timestampTz('created_at')
                 ->default(DB::raw('CURRENT_TIMESTAMP'));
             $table->timestampTz('updated_at')
                 ->default(DB::raw('CURRENT_TIMESTAMP'));
-            $table->unsignedInteger('client_id')->index();
+            $table->unsignedBigInteger('client_id')->index();
         });
 
         \App\Libraries\Helper\DatabaseLibrary::setUpdatedAtTrigger('oauth_personal_access_clients');
